Calculator page UI (#5)
* feat: Update Hero component animation timing

The Hero component animation timing has been adjusted to improve the user experience. The opacity transition now starts slightly later, with a delay of 2.9 seconds instead of 2.8 seconds. Additionally, the call-to-action element now animates with a delay of 2.9 seconds instead of 2.8 seconds. This change enhances the visual flow and timing of the Hero component.

* feat: Add lazy loading to CalculatorPage component

* feat: Add @radix-ui/react-separator package for CalculatorPage component

* feat: Add lazy loading to CalculatorPage component

* feat: Update component classNames for responsive layout

The classNames of several components have been updated to improve the responsiveness of the layout. Specifically, the `IncomeInput`, `PayCycleSelect`, and `DeductionsInput` components now have modified classNames to adjust their width on larger screens. This change ensures a better user experience across different devices and screen sizes.

* feat: Improve responsiveness of layout components

Update classNames of several components to enhance the responsiveness of the layout. Specifically, modify classNames for `IncomeInput`, `PayCycleSelect`, and `DeductionsInput` components to adjust their width on larger screens. This change ensures a better user experience across different devices and screen sizes.

* feat: Improve responsiveness of layout components

Update classNames of several components to enhance the responsiveness of the layout. Specifically, modify classNames for `IncomeInput`, `PayCycleSelect`, and `DeductionsInput` components to adjust their width on larger screens. This change ensures a better user experience across different devices and screen sizes.

* feat: Improve responsiveness of layout components

* feat: Update classNames for responsive layout

Adjust classNames for `IncomeInput`, `PayCycleSelect`, and `DeductionsInput` components to improve responsiveness on larger screens. This ensures a better user experience across different devices and screen sizes.

* feat: Add prettier formatting script to package.json

* feat: Improve Hero component animation timing

Adjust the timing of the Hero component animation to enhance the user experience. The opacity transition now starts slightly later, with a delay of 2.9 seconds instead of 2.8 seconds. Additionally, the call-to-action element now animates with a delay of 2.9 seconds instead of 2.8 seconds. This change improves the visual flow and timing of the Hero component.

* feat: Update CalculatorPage component with text-amber-500 class

Apply the `text-amber-500` class to the SVG icon in the CalculatorPage component to change its color. This improves the visual appearance and consistency of the component.