Add auth (#6)
* feat: Add next-auth package for authentication

Add the next-auth package to enable authentication functionality in the application. This package provides a simple and secure way to handle user authentication, including features like session management, OAuth providers, and JWT token support. Integrating next-auth will enhance the security and user experience of the application.

* npm uninstall

* feat: Add lazy loading for images in next.config

Enable lazy loading for images in the next.config file by adding a remotePatterns property. This property specifies the hostname of the images that should be lazily loaded. This change improves the page load performance by deferring the loading of images until they are needed.