Feat add toast and swr (#11)
* feat: Remove unused fetchRecords function in CalculatorPage component

* feat: Update npm dependencies to include swr@2.2.5

This commit updates the npm dependencies in the package.json file to include swr@2.2.5. This version of swr provides improved functionality and bug fixes, enhancing the performance and reliability of the application.

* feat: Improve loading message in MyIncomeRecords component

This commit updates the loading message in the MyIncomeRecords component to provide a better user experience. Instead of displaying a plain "Loading..." text, it now uses a centered text style with the class "text-center". This change enhances the visual presentation of the loading message.

* feat: Update loading message in MyIncomeRecords component

This commit improves the loading message in the MyIncomeRecords component by using a centered text style with the class "text-center" instead of a plain "Loading..." text. This change enhances the visual presentation of the loading message, providing a better user experience.

* feat: Refactor CalculatorPage component to use useSaveIncomeRecord hook

This commit refactors the CalculatorPage component to use the useSaveIncomeRecord hook instead of manually handling the save functionality. This change improves code readability and maintainability by abstracting the save logic into a reusable hook. It also removes the need for the session check and error handling, as they are now handled within the hook itself.