chore: Refactor lazy loading for images in CalculatorPageWrapper and … (#17)
* chore: Refactor lazy loading for images in CalculatorPageWrapper and Page components

* Refactor lazy loading for images in CalculatorPageWrapper and Page components

* feat: Refactor DeductionsInput component to use formatted currency hook

The DeductionsInput component has been refactored to use a custom hook called useFormattedCurrency. This hook takes an atom value as input and returns the formatted currency value and a setter function for the raw value. This refactoring improves code readability and reusability by abstracting the currency formatting logic into a separate hook.

* Refactor DeductionsInput and IncomeInput components to use useCurrencyFormatter hook

The DeductionsInput and IncomeInput components have been refactored to use a custom hook called useCurrencyFormatter. This hook takes an atom value as input and returns the formatted currency value and a setter function for the raw value. This refactoring improves code readability and reusability by abstracting the currency formatting logic into a separate hook.

* Refactor PayCycleSelect component to improve code quality and UI alignment

* Refactor PayCycleSelect component to improve code quality and UI alignment

* Refactor PayCycleSelect component to improve UI alignment and code quality

* Refactor IncomeTypeRadioGroup component to add conditional border color

The IncomeTypeRadioGroup component has been refactored to conditionally add a border color to the selected radio button. This improves the visual representation of the selected option and enhances the user experience.